+++ lib/src/operators/index.dart
@@ -1,3 +1,4 @@
+export 'as_nullable.dart';
 export 'debug.dart';
 export 'distinct_by.dart';
 export 'distinct_unique_by.dart';